+++ src/HTMLConverter.php
@@ -30,6 +30,8 @@ class HTMLConverter
     /** table id attribute value. */
     protected string $id_value = '';
     protected XMLConverter $xml_converter;
+    /** @var ?callable(array, array-key): array */
+    protected mixed $formatter = null;
 
     public static function create(): self
     {
@@ -60,6 +62,10 @@ class HTMLConverter
      */
     public function convert(iterable $records, array $header_record = [], array $footer_record = []): string
     {
+        if (null !== $this->formatter) {
+            $records = MapIterator::fromIterable($records, $this->formatter);
+        }
+
         $doc = new DOMDocument('1.0');
         if ([] === $header_record && [] === $footer_record) {
             $table = $this->xml_converter->import($records, $doc);
@@ -156,6 +162,19 @@ class HTMLConverter
     {
         $clone = clone $this;
         $clone->xml_converter = $this->xml_converter->fieldElement('td', $fieldname_attribute_name);
+
+        return $clone;
+    }
+
+    /**
+     * Set a callback to format each item before json encode.
+     *
+     * @param ?callable(array, array-key): array $formatter
+     */
+    public function formatter(?callable $formatter): self
+    {
+        $clone = clone $this;
+        $clone->formatter = $formatter;
 
         return $clone;
     }