+++ src/XMLConverter.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace League\Csv;
 
-use Deprecated;
 use DOMAttr;
 use DOMDocument;
 use DOMElement;
@@ -35,6 +34,8 @@ class XMLConverter
     protected string $column_attr = '';
     /** XML offset attribute name. */
     protected string $offset_attr = '';
+    /** @var ?callable(array, array-key): array */
+    public mixed $formatter = null;
 
     public static function create(): self
     {
@@ -56,6 +57,10 @@ class XMLConverter
      */
     public function convert(iterable $records): DOMDocument
     {
+        if (null !== $this->formatter) {
+            $records = MapIterator::fromIterable($records, $this->formatter);
+        }
+
         $doc = new DOMDocument('1.0');
         $node = $this->import($records, $doc);
         $doc->appendChild($node);
@@ -144,6 +149,19 @@ class XMLConverter
     {
         $clone = clone $this;
         $clone->root_name = $this->filterElementName($node_name);
+
+        return $clone;
+    }
+
+    /**
+     * Set a callback to format each item before json encode.
+     *
+     * @param ?callable(array, array-key): array $formatter
+     */
+    public function formatter(?callable $formatter): self
+    {
+        $clone = clone $this;
+        $clone->formatter = $formatter;
 
         return $clone;
     }